backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cymem==2.0.6
defusedxml==0.7.1
Django==4.0.6
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
ginza==5.1.1
idna==3.3
itypes==1.2.0
ja-ginza==5.1.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pathy==0.6.2
plac==1.3.5
preshed==3.0.6
pycparser==2.21
pydantic==1.8.2
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
smart-open==5.2.1
social-auth-app-django==4.0.0
social-auth-core==4.3.0
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.3
sqlparse==0.4.2
srsly==2.4.4
SudachiDict-core==20220519
SudachiPy==0.6.6
thinc==8.0.17
tqdm==4.64.0
typer==0.4.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
wasabi==0.10.1
